receipt-detector/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow>=10.0.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
opencv-python-headless==4.9.0.80
pydantic==2.5.3
pydantic-settings==2.1.0
numpy>=1.24.0,<2.0.0